requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
aniso8601==9.0.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
base58==2.1.1
bitarray==2.6.0
black==23.3.0
blinker==1.6.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.3
cytoolz==0.12.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==1.2.0
Flask==2.3.2
flask-restplus==0.13.0
flask-restx==1.1.0
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.4
importlib-metadata==6.6.0
ipfshttpclient==0.8.0a2
ipykernel==6.22.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.16.0
jupyter_client==8.2.0
jupyter_core==5.3.0
lru-dict==1.1.8
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==1.0.0
nest-asyncio==1.5.6
netaddr==0.8.0
networkx==2.8.6
packaging==23.1
parsimonious==0.8.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==3.20.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.15.0
Pygments==2.15.1
pymongo==4.3.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
pyzmq==25.0.2
requests==2.28.1
rlp==2.0.1
schedule==1.2.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.12
varint==1.0.2
wcwidth==0.2.6
web3==5.31.0
websockets==9.1
Werkzeug==2.3.3
yarl==1.8.1
zipp==3.15.0
